medis_manager/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-heroku==0.3.1
django-jsonfield==1.4.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
